requirements.txt
certifi==2018.4.16
chardet==3.0.4
DateTime==4.2
idna==2.7
pytz==2018.5
requests==2.19.1
urllib3==1.23
zope.interface==4.5.0
python-dotenv==0.8.2
